+++ openidconnect_web/lib/openidconnect_web.dart
@@ -28,8 +28,8 @@ class OpenIdConnectWeb extends OpenIdConnectPlatform {
   }) async {
     if (useWebRedirectLoop) {
       const AUTH_DESTINATION_KEY = "openidconnect_auth_destination_url";
-      html.window.sessionStorage[AUTH_DESTINATION_KEY] =
-          html.window.location.toString();
+      html.window.sessionStorage
+          .setItem(AUTH_DESTINATION_KEY, html.window.location.toString());
       html.window.location.assign(authorizationUrl);
       return Future<String?>.value(null);
     }
@@ -62,7 +62,7 @@ class OpenIdConnectWeb extends OpenIdConnectPlatform {
   Future<String?> processStartup() async {
     const AUTH_RESPONSE_KEY = "openidconnect_auth_response_info";
 
-    final url = html.window.sessionStorage[AUTH_RESPONSE_KEY];
+    final url = html.window.sessionStorage.getItem(AUTH_RESPONSE_KEY);
     html.window.sessionStorage.removeItem(AUTH_RESPONSE_KEY);
 
     return url;